src/app.py
import os
from azure.iot.hub import IoTHubRegistryManager
from azure.ai.personalizer import PersonalizerClient
from azure.identity import DefaultAzureCredential
from azure.mgmt.synapse import SynapseManagementClient

# Azure credentials and configuration
subscription_id = os.getenv("AZURE_SUBSCRIPTION_ID")
resource_group = "SmartFinAdvisorRG"
iot_hub_name = "SmartFinAdvisorIoTHub"
personalizer_endpoint = os.getenv("PERSONALIZER_ENDPOINT")
personalizer_api_key = os.getenv("PERSONALIZER_API_KEY")

# IoT Hub Registry Manager
iot_registry_manager = IoTHubRegistryManager.from_connection_string(os.getenv("IOT_HUB_CONNECTION_STRING"))

# Personalizer Client
personalizer_client = PersonalizerClient(endpoint=personalizer_endpoint, credential=DefaultAzureCredential())

# Synapse Management Client
synapse_client = SynapseManagementClient(credential=DefaultAzureCredential(), subscription_id=subscription_id)

def analyze_financial_data(user_data):
    # Placeholder function to analyze financial data
    # Implement your data analysis logic here
    pass

def get_personalized_advice(user_data):
    # Placeholder function to get personalized financial advice
    # Implement your Personalizer logic here
    pass

def track_spending(device_id):
    # Placeholder function to track spending using IoT Hub
    # Implement your IoT Hub logic here
    pass

def main():
    user_data = {
        "income": 50000,
        "expenses": 20000,
        "savings": 10000
    }
    analyze_financial_data(user_data)
    advice = get_personalized_advice(user_data)
    print("Personalized Financial Advice:", advice)
    track_spending("device123")

if __name__ == "__main__":
    main()
