src/requirements.txt
sql
Copy code
azure-iot-hub
azure-ai-personalizer
azure-identity
azure-mgmt-synapse
